refactor(scraper): removing the old spider with playwright and renaming the new one with API endpoint
