+++ scraper/samsung/spiders/samsung.py
@@ -1,148 +1,138 @@
 import scrapy
+import json
+
+from datetime import datetime
+from dataclasses import dataclass, asdict
+from typing import Literal
+
+
+CLASSIFICATION_NUMBERS = (
+    33010000, 33020000, 39120000, 33110000, 33120000, 33050000, 34060000, 34080000,
+    41040000, 41090000, 34090000, 36010000, 36020000, 36080000, 36070000, 36030000,
+    36060000, 37020000, 37080000, 37090000, 37010000, 37040000, 37030000, 37050000,
+    37070000, 39020000, 39060000, 39030000, 39040000, 39070000, 40030000,
+    100020195, 100034611, 100027453, 100037843, 100020234, 100024278,
+)
+URL_ENDPOINT = "https://www.samsung.com/sec/pf/goodsList"
+URL_PREFIX = "https://www.samsung.com/sec/"
+DATETIME_FORMAT = "%Y/%d/%m %H:%M:%S"
+
+
+@dataclass(eq=False)
+class URLEndpointParameters:
+    dispClsfNo: str
+    pfFasterUseYn: Literal["Y", "N"] = "Y"
+    page: int = 1
+    rows: int = 25535
+    secApp: bool = False
+    ehcacheYn: Literal["Y", "N"] = "Y"
+    soldOutExceptYn: Literal["Y", "N"] = "N"
 
 
 class SamsungSpider(scrapy.Spider):
     name = "samsung"
     allowed_domains = ["www.samsung.com"]
-    start_urls = [
-	    # "https://www.samsung.com/sec/smartphones/all-smartphones/",
-	    # "https://www.samsung.com/sec/tablets/all-tablets/",
-	    # "https://www.samsung.com/sec/galaxybook/all-galaxybook/",
-	    # "https://www.samsung.com/sec/watches/all-watches/",
-	    # "https://www.samsung.com/sec/buds/all-buds/",
-	    # "https://www.samsung.com/sec/mobile-accessories/all-mobile-accessories/",
-	    # "https://www.samsung.com/sec/tvs/all-tvs/",
-	    # "https://www.samsung.com/sec/lifestyletv/all-lifestyletv/",
-	    # "https://www.samsung.com/sec/samsung-audio/all-samsung-audio/",
-	    # "https://www.samsung.com/sec/harman-life-style-audio/all-harman-life-style-audio/",
-	    # "https://www.samsung.com/sec/tv-accessories/all-tv-accessories/",
-	    # "https://www.samsung.com/sec/refrigerators/all-refrigerators/",
-	    # "https://www.samsung.com/sec/kimchi-refrigerators/all-kimchi-refrigerators/",
-	    # "https://www.samsung.com/sec/dishwashers/all-dishwashers/",
-	    # "https://www.samsung.com/sec/water-purifier/all-water-purifier/",
-	    # "https://www.samsung.com/sec/electric-range/all-electric-range/",
-	    # "https://www.samsung.com/sec/cooking-appliances/all-cooking-appliances/",
-	    # "https://www.samsung.com/sec/micro-wave-ovens/all-micro-wave-ovens/",
-	    # "https://www.samsung.com/sec/hood/all-hood/",
-	    # "https://www.samsung.com/sec/kitchen-small-appliance/all-kitchen-small-appliance/",
-	    # "https://www.samsung.com/sec/kitchen-accessories/all-kitchen-accessories/",
-	    # "https://www.samsung.com/sec/washing-machines/all-washing-machines/",
-	    "https://www.samsung.com/sec/dryers/all-dryers/",
-	    # "https://www.samsung.com/sec/airdresser/all-airdresser/",
-	    # "https://www.samsung.com/sec/shoedresser/all-shoedresser/",
-	    # "https://www.samsung.com/sec/air-conditioners/all-air-conditioners/",
-	    # "https://www.samsung.com/sec/system-air-conditioners/all-system-air-conditioners/",
-	    # "https://www.samsung.com/sec/air-cleaner/all-air-cleaner/",
-	    # "https://www.samsung.com/sec/air-cleaner/all-air-cleaner/",
-	    # "https://www.samsung.com/sec/small-appliances/all-small-appliances/",
-	    # "https://www.samsung.com/sec/living-accessories/all-living-accessories/",
-	    # "https://www.samsung.com/sec/desktop/all-desktop/",
-	    # "https://www.samsung.com/sec/pc-accessories/all-pc-accessories/",
-	    # "https://www.samsung.com/sec/monitors/all-monitors/",
-	    # "https://www.samsung.com/sec/memory-storage/all-memory-storage/",
-	    # "https://www.samsung.com/sec/memory-storage/all-memory-storage/",
-	    # "https://www.samsung.com/sec/printer-supplies/all-printer-supplies/",
-	    # "https://www.samsung.com/sec/smartthings-accessories/all-smartthings-accessories/",
-    ]
 
     def start_requests(self):
-        for products_list_url in self.start_urls:
+        from scrapy.utils.url import add_or_replace_parameters
+
+        for classification_number in CLASSIFICATION_NUMBERS:
+            parameters = URLEndpointParameters(dispClsfNo=classification_number)
+            url = add_or_replace_parameters(
+                url=URL_ENDPOINT,
+                new_parameters=asdict(parameters),
+            )
+
             yield scrapy.Request(
-                url=products_list_url,
-                meta={
-                    "playwright": True,
-                    "playwright_include_page": True,
-                    "playwright_context": "product_lists",
-                },
+                url=url,
                 callback=self.parse,
             )
-       
-    async def parse(self, response: scrapy.http.Response):
-        cards_urls = []
-        page = response.meta["playwright_page"]
 
-        await page.wait_for_selector("button#morePrd")
+    def parse(self, response: scrapy.http.Response):
+        response_json = json.loads(response.body)
+        self.log(response_json)
 
-        amount_of_sets = int(await page.locator("span#totalPageCount").text_content())
+        for item in response_json["products"]:
+            has_outlet = False
+            prices = None
+            additional_properties = None
 
-        item_number = 11
-        for _ in range(amount_of_sets):
-            await page.wait_for_selector(f"li.item:nth-child({item_number})")
-            await page.evaluate("morePrd()")
-            item_number += 12
+            if item.get("activatePhoneYn") == "Y":
+                prices = get_prices_schema()
+            if item.get("outletFlgYn") == "Y":
+                has_outlet = True
+            if prices is None:
+                prices = deserialize_price_string(item.get("priceStr"), has_outlet=has_outlet)
 
-        for card_url_locator in await page.locator("li.item a.link-review").all():
-            card_url_postfix = await card_url_locator.get_attribute("href")
-            card_url_postfix = card_url_postfix.rstrip("?focus=review")
-            card_url = response.urljoin(card_url_postfix)
-            cards_urls.append(card_url)
-
-        for n, card_url in enumerate(cards_urls, start=1):
-            context_number = n % amount_of_sets
-            yield scrapy.Request(
-                url=card_url,
-                meta={ "playwright": True, "playwright_context": f"product_items_{context_number}" },
-                callback=self.parse_card,
+            addtional_properties = deserialize_additional_properties_string(
+                item.get("goodsOptStr")
             )
 
-    async def parse_card(self, response: scrapy.http.Response):
-        prices_selector = response.css("div.itm-price")
-        props_selector = response.css("div.itm-option-choice")
+            yield {
+                "title": item.get("goodsNm"),
+                "model": item.get("mdlNm"),
+                "model_code": item.get("mdlCode"),
+                "link": URL_PREFIX + item.get("goodsDetailUrl"),
+                "item_category": item.get("compDispClsfEnNm"),
+                "item_classification_number": item.get("compDispClsfNo"),
+                "rating": item.get("reviewGrade"),
+                "reviews_quantity": item.get("reviewCount"),
+                "stock_quantity": item.get("stockQty"), 
+                "datetime": datetime.now().strftime(DATETIME_FORMAT),
+                "coupon_discount": item.get("cpAllDcAmt"),
+                "additional_properties": addtional_properties,
+                **prices,
+            }
 
-        return {
-            "title": response.css("h1#goodsDetailNm::text").get(),
-            "url": response.url,
-            "model": response.css("div.itm-sku::text").get(),
-            "category": self.name,
 
-            # "benefit_price_validity_period": response.css().get(),
-            # "coupon_discount_validity_period": response.css().get(),
+def get_prices_schema():
+    return {
+        "standard_price": None,
+        "member_price": None,
+        "benefit_price": None,
+        "outlet_special_price": None,
+    }
 
-            "rating": response.css("div.itm-sart-rating span::text").get(),
-            "quantity_of_reviews": response.css("a.itm-review-count::text").get(),
-            **self.parse_prices(prices_selector),
-            **self.parse_additional_props(props_selector),
-        }
 
-    def parse_prices(self, prices_selector: scrapy.Selector):
-        prices = {
-            "standard_price": None,
-            "member_price": None,
-            "benefit_price": None,
-            "outlet_special_price": None,
-            "coupon_discount_quantity": None,
-            "coupon_discounted_price": None,
-        }
-        extract_quantity = lambda pb: pb.css("dd span::text").get().strip()
+def deserialize_price_string(price_split: str, has_outlet: bool):
+    price_split = price_split.split("|")
+    prices_schema = get_prices_schema()
 
-        for price_selector in prices_selector.css("dl"):
-            self.log(price_selector.css("dt::text").get().strip())
-            match price_selector.css("dt::text").get().strip():
-                case "기준가":
-                    prices["standard_price"] = extract_quantity(price_selector)
-                case "회원가":
-                    prices["member_price"] = extract_quantity(price_selector)
-                case "혜택가":
-                    prices["benefit_price"] = extract_quantity(price_selector)
-                case "아울렛 특가":
-                    prices["outlet_special_price"] = extract_quantity(price_selector)
-                case "쿠폰 할인 금액":
-                    prices["coupon_discount_quantity"] = extract_quantity(price_selector)
-                case "쿠폰 적용 예상가":
-                    prices["coupon_discounted_price"] = extract_quantity(price_selector)
-                case other:
-                    self.logger.warn(f"'{other}' is not defined by the parser")
+    price_split[2], price_split[3] = int(price_split[2]), int(price_split[3])
+    prices_schema["standard_price"] = price_split[2]
 
-        return prices
+    if has_outlet:
+        prices_schema["outlet_special_price"] = int(price_split[-1])
+    elif price_split[1] == '00':
+        prices_schema["member_price"] = price_split[3] if price_split[2] != price_split[3] else None
+    elif price_split[1] != '00':
+        price_split[4] = int(price_split[4])
 
-    def parse_additional_props(self, props_selector: scrapy.Selector):
-        props = {}
+        if price_split[2] == price_split[3] and price_split[3] != price_split[4]:
+            prices_schema["benefit_price"] = price_split[4]
+        elif price_split[2] != price_split[3] and price_split[3] != price_split[4]:
+            prices_schema["member_price"] = price_split[3]
+            prices_schema["benefit_price"] = price_split[4]
 
-        for prop_selector in props_selector.css("dl")[:-1]:
-            if prop_selector.attrib["class"] not in ("count-show-box", "itm-option-etc"):
-                prop_title = prop_selector.css("dt span::text").get()
-                prop_values = prop_selector.css("dd li span::text").getall()
 
-                props[prop_title] = prop_values
+    return prices_schema
 
-        return props
+
+def deserialize_additional_properties_string(additional_properties_str: str):
+    if additional_properties_str is None:
+        return additional_properties_str
+
+    additional_properties_split = additional_properties_str.split("\n")
+    additional_properties_schema = {}
+
+    for property in additional_properties_split:
+        property_split = property.split("|")
+        property_key = property_split[-3]
+        property_value = property_split[-2]
+
+        if additional_properties_schema.get(property_key):
+            additional_properties_schema[property_key].append(property_value)
+        else:
+            additional_properties_schema[property_key] = [property_value]
+
+    return additional_properties_schema
